Port from MESS, nw

+++ src/mame/includes/stv.h
@@ -29,7 +29,7 @@ public:
 		UINT32    dst_add[3];	/* Destination Addition for DMA lv n*/
 		INT32     size[3];		/* Transfer DMA size lv n*/
 		UINT32    index[3];
-		UINT8     start_factor[3];
+		int       start_factor[3];
 		UINT8     enable_mask[3];
 	}m_scu;
